requirements.txt
numpy
pandas
scikit-learn
xgboost
lightgbm
scipy
matplotlib
seaborn
imblearn
mlxtend
shap
sklearn-genetic
#-e .